pyproject.toml
[tool.poetry]
name = "cognigpt"
version = "0.1.0"
description = "GPT-powered multi-agent system inspired by the Global Workspace Theory"
authors = ["Hui Song"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "1.0.0"
pyyaml = "^6.0.1"
strenum = "^0.4.15"
dpkt = "^1.9.8"
paho-mqtt = "^1.6.1"
termcolor = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
